Fix action box a bit
