+++ Assets/_GameJamCat/Prefabs/UserInterface/ActionBoxCanvas.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9dcf679b4dfcc184598888c9e172c2cc
+guid: b7c9554bb8e98724995695ce09b36e78
 PrefabImporter:
   externalObjects: {}
   userData:
