+++ Assets/_GameJamCat/Scripts/UI/ActionBoxMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e0297c2b4072be744aa249e338e35b55
+guid: ce40a4cec5ab02f46994d28478f6ca5f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
